get Lead inputs by console method
