+++ src/main/java/com/ironhack/crm/utils/UtilsUserInputs.java
@@ -1,0 +1,80 @@
+package com.ironhack.crm.utils;
+
+import com.ironhack.crm.domain.models.Lead;
+
+import java.util.Scanner;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class UtilsUserInputs {
+
+
+    public static Lead getUserLeadInput(){
+        return new Lead(getLeadNameInput(), getCompanyNameInput(), getLeadEmailInput(), getLeadPhoneNumberInput());
+    }
+
+
+    public static boolean validateEmail(String email){
+        String regex = "^[_A-Za-z0-9-\\+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";
+        boolean validate = false;
+        Pattern p = Pattern.compile(regex);
+        Matcher m = p.matcher(email);
+        if(m.matches()){
+            validate=true;
+        }
+        return validate;
+    }
+
+    private static String getLeadNameInput(){
+        boolean isName = false;
+        String name = "";
+        Scanner input = new Scanner(System.in);
+        System.out.print("Please input lead's 'FirstName MiddleName  LastName' format: ");
+        while (!isName) {
+            name = input.nextLine();
+            isName = (name.trim().split("\\s+").length == 3);
+            if (!isName) {
+                System.out.print("\nEnter your name as 'FirstName MiddleName  LastName' format: ");
+            }
+        }
+        return name;
+    }
+
+    private static String getCompanyNameInput(){
+        String companyName = "";
+        Scanner input = new Scanner(System.in);
+        System.out.print("Please input the lead's company name: ");
+        companyName = input.nextLine();
+        return companyName;
+    }
+
+    private static String getLeadEmailInput(){
+        boolean isEmail = false;
+        String email = "";
+        Scanner input = new Scanner(System.in);
+        System.out.print("Please input lead's email: ");
+        while (!isEmail) {
+            email = input.nextLine();
+            isEmail = validateEmail(email);
+            if (!isEmail) {
+                System.out.print("\nPlease input lead's email with correct format (name@domain.example): ");
+            }
+        }
+        return email;
+    }
+
+    private static String getLeadPhoneNumberInput(){
+        boolean isPhoneNumber = false;
+        String phoneNumber = "";
+        Scanner input = new Scanner(System.in);
+        while (!isPhoneNumber) {
+            System.out.print("Please input lead's phone number: ");
+            phoneNumber = input.nextLine();
+            if (phoneNumber.isEmpty()==false) isPhoneNumber = true;
+            if (!isPhoneNumber) {
+                System.out.print("\nPlease input phone number: ");
+            }
+        }
+        return phoneNumber;
+    }
+}
